+++ kubejs/server_scripts/Create Recipes/greg.js
@@ -1,40 +1,5 @@
 ServerEvents.recipes(event => {
-//Casings
-    function casing(input, output, circuit){
-        event.recipes.create.mechanical_crafting(output, [
-            'AAA',
-            'ABA',
-            'AAA'
-        ], {
-            A: input,
-            B: circuit
-        })
-    }
-//Bricked Hulls
-    function hull(input, tempitem, output, add_one, dust){
-        event.recipes.create.sequenced_assembly([
-            Item.of(output).withChance(95.0),
-            Item.of('2x gtceu:small_brick_dust').withChance(3.0),
-            Item.of(dust).withChance(2.0)
-        ], input, [
-        event.recipes.createDeploying(tempitem, [tempitem, add_one]),
-        //event.recipes.vintageimprovements(tempitem, [tempitem]),
-        event.recipes.createPressing(tempitem, [tempitem])
-        ]).transitionalItem(tempitem).loops(5)
-    }
 //Steam machines
-    function solidboiler(output, plate, burner, hull){
-        event.recipes.create.mechanical_crafting(output, [
-            'AAA',
-            'ABA',
-            'CDC'
-        ], {
-            A: plate,
-            B: burner,
-            C: 'minecraft:bricks',
-            D: hull
-        })
-    }
     function solarboiler(output, plate, pipe, hull){
         event.recipes.create.mechanical_crafting(output, [
             'AAAAA',
@@ -49,6 +14,36 @@ ServerEvents.recipes(event => {
             E: hull
         })
     }
+//Poor steel
+    event.recipes.create.mixing(
+        ['2x gtceu:poor_steel_ingot'], 
+        ['2x minecraft:iron_ingot', '2x minecraft:coal']
+        ).heatRequirement('heated')
+    event.recipes.create.pressing(
+        'gtceu:wrought_iron',
+        'gtceu:poor_steel'
+    )
+    event.replaceInput(
+        { input: 'tfmg:cast_iron_ingot' },
+        'tfmg:cast_iron_ingot',
+        'gtceu:poor_steel_ingot'
+        )
+    event.replaceInput(
+        { input: 'createdeco:industrial_iron_ingot' },
+        'createdeco:industrial_iron_ingot',
+        'gtceu:poor_steel_ingot'
+        )
+    event.replaceInput(
+        { input: 'createdeco:industrial_iron_sheet' },
+        'createdeco:industrial_iron_sheet',
+        'gtceu:poor_steel_plate'
+        )
+    event.replaceInput(
+        { input: 'createdeco:zinc_sheet' },
+        'createdeco:zinc_sheet',
+        'gtceu:zinc_plate'
+        )
+    
 //Concrete
     event.recipes.create.mixing(
         [Fluid.of(('gtceu:concrete'), 1000)], 
@@ -59,6 +54,27 @@ ServerEvents.recipes(event => {
         ['6x gtceu:firebrick', '2x gtceu:gypsum_dust', Fluid.of(('gtceu:concrete'), 750)]
         ).heatRequirement('heated')
 //Functions
+    const bolts = Ingredient.of('#forge:bolts').itemIds
+    bolts.forEach( (base) => {
+        event.custom({
+            type: 'create:cutting',
+            ingredients: [
+                {
+                    item: base.slice(0, -5) + '_rod'
+                }
+            ],
+            results: [
+                {
+                    item: base,
+                    count: 2
+                }
+            ],
+            processingTime: 50
+        })
+        event.recipes.gtceu.assembler(base.slice(0, -4) + '_rod')
+            .itemInputs(base)
+            .itemOutputs(base)
+    })
     function milling(output, input){
         event.recipes.create.milling(output, input)
     }
@@ -69,14 +85,6 @@ ServerEvents.recipes(event => {
         event.recipes.create.filling(output, input)
     }
 
-    casing('gtceu:bronze_plate', 'gtceu:bronze_machine_casing', 'kubejs:clockwork_mechanism')
-    casing('gtceu:steel_plate', 'gtceu:steel_machine_casing', 'kubejs:advanced_clockwork_mechanism')
-    hull('minecraft:bricks', 'kubejs:incomplete_bronze_hull', 'gtceu:bronze_brick_casing', 'gtceu:bronze_plate', '2x gtceu:small_bronze_dust')
-    hull('minecraft:bricks', 'kubejs:incomplete_steel_hull', 'gtceu:steel_brick_casing', 'gtceu:wrought_iron_plate', '2x gtceu:small_wrought_iron_dust')
-    solidboiler('gtceu:lp_steam_solid_boiler', 'gtceu:bronze_plate', 'minecraft:furnace', 'gtceu:bronze_brick_casing')
-    solidboiler('gtceu:hp_steam_solid_boiler', 'gtceu:steel_plate', 'minecraft:furnace', 'gtceu:steel_brick_casing')
-    solidboiler('gtceu:lp_steam_liquid_boiler', 'gtceu:bronze_plate', 'fluidtank:tank_wood', 'gtceu:bronze_brick_casing')
-    solidboiler('gtceu:hp_steam_liquid_boiler', 'gtceu:steel_plate', 'fluidtank:tank_iron', 'gtceu:steel_brick_casing')
     solarboiler('gtceu:lp_steam_solar_boiler', 'gtceu:bronze_plate', 'gtceu:bronze_small_fluid_pipe', 'gtceu:bronze_brick_casing')
     solarboiler('gtceu:hp_steam_solar_boiler', 'gtceu:steel_plate', 'gtceu:steel_small_fluid_pipe', 'gtceu:steel_brick_casing')
     milling('gtceu:clay_dust', 'minecraft:clay_ball')