+++ kubejs/startup_scripts/gtceu/aaa Materials and Elements/aaa_icon_sets.js
@@ -1,4 +1,6 @@
 GTCEuStartupEvents.registry('gtceu:material_icon_set', event => {
     event.create('infinity')
         .parent('shiny')
+    event.create('stellar_matter')
+        .parent('dull')
 })